// app/sanity/fellesKomponenter/header/types.ts
export interface IMeldekortHeader {
  skipLink: string;
  systemHeaderAriaLabel: string;
  homeLink: string;
  homeLinkAriaLabel: string;
  userButtonAriaLabel: string;
  dropdownAriaLabel: string;
  sensitiveDataSwitchLabel: string;
  sensitiveDataSwitchDescription: string;
  logoutLinkText: string;
  darkThemeActive: string;
  lightThemeActive: string;
}